+++ 0x01-python_async_function/4-tasks.py
@@ -5,7 +5,7 @@ from typing import List
 task_wait_random = __import__("3-tasks").task_wait_random
 
 
-async def task_wait_n(n: int, max_delay: float) -> List[float]:
+async def task_wait_n(n: int, max_delay: int) -> List[float]:
     """Create coroutines and await their completion upon which
     they will be added to the list to be returned"""
     tasks = [task_wait_random(max_delay) for _ in range(n)]
